+++ src/hash.h
@@ -26,15 +26,15 @@ typedef struct {
   size_t bloom;
 } HashTable;
 
-TableEntry *table_lookup(HashTable *tbl, const char *key_ptr, int key_len) __attribute__ ((pure));
+TableEntry *table_lookup(HashTable *tbl, const char *key_ptr, size_t key_len) __attribute__ ((pure));
 
-TableEntry *table_lookup_with_hash(HashTable *tbl, const char *key_ptr, int key_len, size_t key_hash) __attribute__ ((pure));
+TableEntry *table_lookup_with_hash(HashTable *tbl, const char *key_ptr, size_t key_len, size_t key_hash) __attribute__ ((pure));
 
 // if the key was not found, return null but allocate a mapping in first_free_ptr
-TableEntry *table_lookup_alloc(HashTable *tbl, const char *key_ptr, int key_len, TableEntry** first_free_ptr);
+TableEntry *table_lookup_alloc(HashTable *tbl, const char *key_ptr, size_t key_len, TableEntry **first_free_ptr);
 
 // added in case you've already precomputed the hash for other reasons, and wanna avoid double computing it
-TableEntry *table_lookup_alloc_with_hash(HashTable *tbl, const char *key_ptr, int key_len, size_t key_hash, TableEntry** first_free_ptr);
+TableEntry *table_lookup_alloc_with_hash(HashTable *tbl, const char *key_ptr, size_t key_len, size_t key_hash, TableEntry** first_free_ptr);
 
 // thanks http://stackoverflow.com/questions/7666509/hash-function-for-string
 static inline size_t hash(const char *ptr, int len) {
